Split out ShellSentry program from CLI
This makes it easier to use and wrap in other programs or CLIs
