+++ src/main/kotlin/slack/cli/shellsentry/ShellSentryCli.kt
@@ -22,20 +22,15 @@ import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.types.path
 import com.squareup.moshi.adapter
-import eu.jrie.jetbrains.kotlinshell.shell.shell
-import java.nio.file.Path
-import kotlin.io.path.ExperimentalPathApi
 import kotlin.io.path.createDirectories
-import kotlin.io.path.createTempFile
-import kotlin.io.path.deleteRecursively
-import kotlin.system.exitProcess
 import okio.buffer
 import okio.source
 import org.jetbrains.annotations.TestOnly
 import slack.cli.projectDirOption
 
 /**
- * Executes a command with Bugsnag tracing and retries as needed.
+ * Executes a command with Bugsnag tracing and retries as needed. This CLI is a shim over
+ * [ShellSentry].
  *
  * Example:
  * ```
@@ -46,8 +41,11 @@ public class ShellSentryCli :
   CliktCommand("Executes a command with Bugsnag tracing and retries as needed.") {
 
   internal val projectDir by projectDirOption()
+
   internal val verbose by option("--verbose", "-v").flag()
+
   internal val bugsnagKey by option("--bugsnag-key", envvar = "PE_BUGSNAG_KEY")
+
   internal val configurationFile by
     option("--config", envvar = "PE_CONFIGURATION_FILE")
       .path(mustExist = true, canBeFile = true, canBeDir = false)
@@ -61,12 +59,11 @@ public class ShellSentryCli :
         help = "Instructs this CLI to not exit the process with the status code. Test only!"
       )
       .flag()
+
   @get:TestOnly internal val parseOnly by option("--parse-only").flag(default = false)
 
   internal val args by argument().multiple()
 
-  @Suppress("CyclomaticComplexMethod", "LongMethod")
-  @OptIn(ExperimentalStdlibApi::class, ExperimentalPathApi::class)
   override fun run() {
     if (parseOnly) return
 
@@ -77,108 +74,22 @@ public class ShellSentryCli :
         it.source().buffer().use { source -> moshi.adapter<ShellSentryConfig>().fromJson(source) }
       }
         ?: ShellSentryConfig()
-    // The command to be executed
-    val cmd = args.joinToString(" ")
 
-    // Temporary file for command output
-    val tmpDir = projectDir.resolve("tmp/shellsentry")
-    tmpDir.createDirectories()
+    // Temporary dir for command output
+    val cacheDir = projectDir.resolve("tmp/shellsentry")
+    cacheDir.createDirectories()
 
-    // Initial command execution
-    val (initialExitCode, initialLogFile) = executeCommand(cmd, tmpDir)
-    var exitCode = initialExitCode
-    var logFile = initialLogFile
-    var attempts = 0
-    while (exitCode != 0 && attempts < 1) {
-      attempts++
-      echo(
-        "Command failed with exit code $exitCode. Running processor script (attempt $attempts)..."
+    ShellSentry(
+        command = args.joinToString(" "),
+        workingDir = projectDir,
+        cacheDir = cacheDir,
+        config = config,
+        verbose = verbose,
+        bugsnagKey = bugsnagKey,
+        debug = debug,
+        noExit = noExit,
+        echo = ::echo
       )
-
-      echo("Processing CI failure")
-      val resultProcessor = ResultProcessor(verbose, bugsnagKey, config, ::echo)
-
-      when (val retrySignal = resultProcessor.process(logFile, false)) {
-        is RetrySignal.Ack,
-        RetrySignal.Unknown -> {
-          echo("Processor exited with 0, exiting with original exit code...")
-          break
-        }
-        is RetrySignal.RetryDelayed -> {
-          echo(
-            "Processor script exited with 2, rerunning the command after ${retrySignal.delay}..."
-          )
-          // TODO add option to reclaim memory?
-          Thread.sleep(retrySignal.delay.inWholeMilliseconds)
-          val secondResult = executeCommand(cmd, tmpDir)
-          exitCode = secondResult.exitCode
-          logFile = secondResult.outputFile
-          if (secondResult.exitCode != 0) {
-            // Process the second failure, then bounce out
-            resultProcessor.process(secondResult.outputFile, isAfterRetry = true)
-          }
-        }
-        is RetrySignal.RetryImmediately -> {
-          echo("Processor script exited with 1, rerunning the command immediately...")
-          // TODO add option to reclaim memory?
-          val secondResult = executeCommand(cmd, tmpDir)
-          exitCode = secondResult.exitCode
-          logFile = secondResult.outputFile
-          if (secondResult.exitCode != 0) {
-            // Process the second failure, then bounce out
-            resultProcessor.process(secondResult.outputFile, isAfterRetry = true)
-          }
-        }
-      }
-    }
-
-    // If we got here, all is well
-    // Delete the tmp files
-    if (!debug) {
-      tmpDir.deleteRecursively()
-    }
-
-    echo("Exiting with code $exitCode")
-    if (!noExit) {
-      exitProcess(exitCode)
-    }
+      .exec()
   }
-
-  // Function to execute command and capture output. Shorthand to the testable top-level function.
-  private fun executeCommand(command: String, tmpDir: Path) =
-    executeCommand(projectDir, command, tmpDir, ::echo)
-}
-
-internal data class ProcessResult(val exitCode: Int, val outputFile: Path)
-
-// Function to execute command and capture output
-internal fun executeCommand(
-  workingDir: Path,
-  command: String,
-  tmpDir: Path,
-  echo: (String) -> Unit,
-): ProcessResult {
-  echo("Running command: '$command'")
-
-  val tmpFile = createTempFile(tmpDir, "shellsentry", ".txt").toAbsolutePath()
-
-  var exitCode = 0
-  shell {
-    // Weird but the only way to set the working dir
-    shell(dir = workingDir.toFile()) {
-      // Read the output of the process and write to both stdout and file
-      // This makes it behave a bit like tee.
-      val echoHandler = stringLambda { line ->
-        // The line always includes a trailing newline, but we don't need that
-        echo(line.removeSuffix("\n"))
-        // Pass the line through unmodified
-        line to ""
-      }
-      val process = command.process() forkErr { it pipe echoHandler pipe tmpFile.toFile() }
-      pipeline { process pipe echoHandler pipe tmpFile.toFile() }.join()
-      exitCode = process.process.pcb.exitCode
-    }
-  }
-
-  return ProcessResult(exitCode, tmpFile)
 }
